reworked resolution slider
Initially, I intended for this slider to represent the arc length of the telescope's field of view. This gave me some issues while trying to implement it, so I changed it to be a normalized field of view slider from 0 to 1 with 0 being a lazer beam and 1 being full 360 degree vision. I have figured out how to iplement my original idea. The slider now controls the arc length of the FOV.